+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Update Signal notification template to include more data
 - Add Profile.deletion_scheduled_deleted field, and UI banner when it's set
 - Add support for specifying MessagingServiceSid when sending SMS and WA messages
+- Update the smtpd management command to use the aiosmtpd library
 
 ### Bug Fixes
 - Fix a race condition when pinging and deleting checks at the same time